requirements.txt
mysql-connector-python==8.0.32
flask==2.3.2